Move SDL files into core/platform/SDL for organisation.
