+++ engine/include/engine/core/CMakeLists.txt
@@ -11,3 +11,4 @@ add_subdirectory(events)
 add_subdirectory(graphics)
 add_subdirectory(math)
 add_subdirectory(utils)
+add_subdirectory(platform)
