+++ engine/src/core/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(graphics)
 add_subdirectory(math)
 add_subdirectory(utils)
+add_subdirectory(platform)
