+++ engine/src/core/platform/CMakeLists.txt
@@ -1,0 +1,13 @@
+set(platformSourceDir ${CMAKE_CURRENT_LIST_DIR})
+
+set(platformSources
+    ${platformSourceDir}/SDL/SDLWindow.cpp
+)
+
+source_group("Source Files\\Core\\Platform" FILES ${platformSources})
+
+set(engineSources
+    ${engineSources}
+    ${platformSources}
+    CACHE INTERNAL "Global Engine Sources"
+)
